aliases:

+++ scripts/restart.sh
@@ -1,3 +1,8 @@
+if [ $# -eq 0 ]; then
+  echo "Error: No arguments provided."
+  exit 1
+fi
+
 echo "restart $1"
 docker stop $1
 docker rm $1
